pdp-django-cours/02-django-modul/1-django-rest-framework/Netflix/requirements.txt
Django~=5.0.2
djangorestframework~=3.15.1